module level docstrings, refactor module names..
