+++ docs/_autosummary/augurpy.cv.rst
@@ -3,25 +3,29 @@
 
 .. automodule:: augurpy.cv
 
+   
+   
+   
 
-
-
-
-
-
+   
+   
    .. rubric:: Functions
 
    .. autosummary::
-
+   
+      ccc_score
       run_cross_validation
       set_scorer
+   
+   
 
+   
+   
+   
 
-
-
-
-
-
+   
+   
+
