+++ docs/_autosummary/augurpy.estimators.rst
@@ -1,7 +1,7 @@
 ﻿augurpy.estimators
 ==================
 
-.. automodule:: augurpy.estimators
+.. automodule:: augurpy.estimator
