+++ docs/_autosummary/augurpy.evaluate.rst
@@ -1,0 +1,32 @@
+﻿augurpy.evaluate
+================
+
+.. automodule:: augurpy.evaluate
+
+   
+   
+   
+
+   
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      average_metrics
+      calculate_auc
+      cross_validate_subsample
+      draw_subsample
+   
+   
+
+   
+   
+   
+
+   
+   
+   
+
+
+
