+++ docs/_autosummary/augurpy.read_load.rst
@@ -3,25 +3,28 @@
 
 .. automodule:: augurpy.read_load
 
+   
+   
+   
 
-
-
-
-
-
+   
+   
    .. rubric:: Functions
 
    .. autosummary::
-
+   
       feature_selection
       load
+   
+   
 
+   
+   
+   
 
-
-
-
-
-
+   
+   
+
