+++ docs/usage.rst
@@ -26,5 +26,5 @@ You can then access the respective modules like:
 
    augurpy.read_load
    augurpy.estimators
-   augurpy.calculate_AUC
+   augurpy.evaluate
    augurpy.cv
